third_party/blink/web_tests/fast/forms/input-pattern-expected.txt

Test the behavior of pattern attribute and pattern DOM property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.pattern is ""
PASS input.getAttribute("pattern") is "foo"
PASS input.pattern is "bar"
PASS input.pattern is "null"
PASS input.getAttribute("pattern") is "null"
PASS input.pattern is "null"
PASS input.pattern is "undefined"
PASS input.getAttribute("pattern") is "undefined"
PASS input.pattern is "undefined"
PASS input.pattern is "256"
PASS input.getAttribute("pattern") is "256"
PASS input.pattern is "256"
PASS successfullyParsed is true

TEST COMPLETE
